EES-7150 Changes in response to PR comments
