+++ src/GovUk.Education.ExploreEducationStatistics.Data.Processor/Program.cs
@@ -60,6 +60,7 @@ var host = new HostBuilder()
                 .AddTransient<ImporterLocationService>()
                 .AddTransient<IImporterMetaService, ImporterMetaService>()
                 .AddTransient<IDataImportService, DataImportService>()
+                .AddTransient<IDataSetMappingService, DataSetMappingService>()
                 .AddTransient<IValidatorService, ValidatorService>()
                 .AddSingleton<IFileTypeService, FileTypeService>()
                 .AddSingleton<IGuidGenerator, SequentialGuidGenerator>()
