+++ src/GovUk.Education.ExploreEducationStatistics.Data.Processor/Services/Interfaces/IDataSetMappingService.cs
@@ -1,0 +1,8 @@
+#nullable enable
+
+namespace GovUk.Education.ExploreEducationStatistics.Data.Processor.Services.Interfaces;
+
+public interface IDataSetMappingService
+{
+    Task CreateInitialDataSetMappingIfReplacement(Guid replacementFileId);
+}
